Moved memory space types to include dir dash/memory
