+++ dash/include/dash/List.h
@@ -5,7 +5,7 @@
 #include <dash/GlobRef.h>
 #include <dash/Team.h>
 #include <dash/Exception.h>
-#include <dash/GlobHeapMem.h>
+#include <dash/memory/GlobHeapMem.h>
 #include <dash/Allocator.h>
 #include <dash/Array.h>
 #include <dash/Meta.h>
