+++ dash/include/dash/Matrix.h
@@ -6,7 +6,7 @@
 #include <dash/Team.h>
 #include <dash/Pattern.h>
 #include <dash/GlobRef.h>
-#include <dash/GlobStaticMem.h>
+#include <dash/memory/GlobStaticMem.h>
 #include <dash/Allocator.h>
 #include <dash/HView.h>
 #include <dash/Meta.h>
